chore: refactor code for pool metadata table
