+++ apps/hubble-stats/data/index.ts
@@ -1,4 +1,4 @@
-export { default as getPoolMetadataData } from './getPoolMetadataData';
+export { default as getPoolMetadataData } from './poolTables/getPoolMetadataTableData';
 export { default as getPoolOverviewCardData } from './getPoolOverviewCardData';
 export { default as getTvlChartData } from './charts/getOverviewTvlChartData';
 export { default as getVolumeChartData } from './charts/getOverviewVolumeChartData';
