+++ apps/hubble-stats/data/poolTables/index.ts
@@ -1,4 +1,5 @@
 export { default as getPoolDepositTableData } from './getPoolDepositTableData';
+export { default as getPoolMetadataTableData } from './getPoolMetadataTableData';
 export { default as getPoolRelayerEarningsTableData } from './getPoolRelayerEarningsTableData';
 export { default as getPoolTransactionsTableData } from './getPoolTransactionsTableData';
 export { default as getPoolTwlTableData } from './getPoolTwlTableData';
